requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
Django==4.2.10
django-cors-headers==2.0.0
django-utils-six==2.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.4.0
google-auth==2.28.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradientai==1.7.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
joblib==1.3.2
keras==2.13.1
libclang==16.0.6
llama-index==0.10.5
llama-index-agent-openai==0.1.1
llama-index-core==0.10.5
llama-index-embeddings-gradient==0.1.1
llama-index-embeddings-openai==0.1.1
llama-index-legacy==0.9.48
llama-index-llms-gradient==0.1.1
llama-index-llms-openai==0.1.2
llama-index-multi-modal-llms-openai==0.1.1
llama-index-program-openai==0.1.1
llama-index-question-gen-openai==0.1.1
llama-index-readers-file==0.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==1.12.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pypdf==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.4
teachable-machine==1.1
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tiktoken==0.6.0
tqdm==4.66.2
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.0
werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
